requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.5.2
SQLAlchemy==2.0.34
psycopg[binary]==3.2.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
pytest==8.3.2
ruff==0.6.9
email-validator==2.1.1
python-multipart==0.0.9
